Comment functionality

+++ django_blog/blog/views.py
@@ -1,45 +1,46 @@
-from django.shortcuts import render, get_object_or_404, redirect
-from django.contrib.auth.decorators import login_required
+from django.shortcuts import render, redirect, get_object_or_404
 from .models import Post, Comment
 from .forms import CommentForm
+from django.contrib.auth.decorators import login_required
 
 def post_detail(request, post_id):
-    post = get_object_or_404(Post, pk=post_id)
-    comments = post.comments.all().order_by('-created_at')
-    form = CommentForm()
+    post = get_object_or_404(Post, id=post_id)
+    comments = post.comments.all()
+    comment_form = CommentForm()
 
-    return render(request, 'blog/post_detail.html', {'post': post, 'comments': comments, 'form': form})
-
-@login_required
-def comment_new(request, pk):
-    post = get_object_or_404(Post, pk=pk)
-    if request.method == 'POST':
-        form = CommentForm(request.POST)
-        if form.is_valid():
-            comment = form.save(commit=False)
+    if request.method == "POST" and request.user.is_authenticated:
+        comment_form = CommentForm(request.POST)
+        if comment_form.is_valid():
+            comment = comment_form.save(commit=False)
             comment.post = post
             comment.author = request.user
             comment.save()
-            return redirect('post_detail', post_id=pk)
-    else:
-        form = CommentForm()
-    return render(request, 'blog/comment_new.html', {'form': form, 'post': post})
+            return redirect('post_detail', post_id=post.id)
+
+    return render(request, 'blog/post_detail.html', {'post': post, 'comments': comments, 'comment_form': comment_form})
 
 @login_required
-def comment_edit(request, comment_id):
-    comment = get_object_or_404(Comment, pk=comment_id, author=request.user)
-    if request.method == 'POST':
+def edit_comment(request, comment_id):
+    comment = get_object_or_404(Comment, id=comment_id)
+    if request.user != comment.author:
+        return redirect('post_detail', post_id=comment.post.id)
+
+    if request.method == "POST":
         form = CommentForm(request.POST, instance=comment)
         if form.is_valid():
             form.save()
             return redirect('post_detail', post_id=comment.post.id)
     else:
         form = CommentForm(instance=comment)
-    return render(request, 'blog/comment_edit.html', {'form': form, 'comment': comment})
+
+    return render(request, 'blog/edit_comment.html', {'form': form, 'comment': comment})
 
 @login_required
-def comment_delete(request, comment_id):
-    comment = get_object_or_404(Comment, pk=comment_id, author=request.user)
-    post_id = comment.post.id
-    comment.delete()
-    return redirect('post_detail', post_id=post_id)
+def delete_comment(request, comment_id):
+    comment = get_object_or_404(Comment, id=comment_id)
+    if request.user == comment.author:
+        post_id = comment.post.id
+        comment.delete()
+        return redirect('post_detail', post_id=post_id)
+    return redirect('post_detail', post_id=comment.post.id)
+
